+++ server/src/main/java/org/elasticsearch/bootstrap/Bootstrap.java
@@ -17,6 +17,7 @@ import org.elasticsearch.common.settings.SecureSettings;
 import org.elasticsearch.core.SuppressForbidden;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.node.NodeValidationException;
+import org.elasticsearch.plugins.PluginsLoader;
 
 import java.io.PrintStream;
 
@@ -41,6 +42,9 @@ class Bootstrap {
 
     // the loaded settings for the node, not valid until after phase 2 of initialization
     private final SetOnce<Environment> nodeEnv = new SetOnce<>();
+
+    // loads information about plugins required for entitlements in phase 2, used by plugins service in phase 3
+    private final SetOnce<PluginsLoader> pluginsLoader = new SetOnce<>();
 
     Bootstrap(PrintStream out, PrintStream err, ServerArgs args) {
         this.out = out;
@@ -70,6 +74,14 @@ class Bootstrap {
 
     Environment environment() {
         return nodeEnv.get();
+    }
+
+    void setPluginsLoader(PluginsLoader pluginsLoader) {
+        this.pluginsLoader.set(pluginsLoader);
+    }
+
+    PluginsLoader pluginsLoader() {
+        return pluginsLoader.get();
     }
 
     void exitWithNodeValidationException(NodeValidationException e) {
